// src/lib/generated/billing/models/discount.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * spaces-billing-and-subscription
 * Spaces Platform billing and subscription service
 * OpenAPI spec version: 1.0.0
 */
import type { DiscountFixedValue } from './discountFixedValue';
import type { DiscountPercentage } from './discountPercentage';

/**
 * Model Discount
 */
export interface Discount {
  /** @nullable */
  modifiedAt: string | null;
  createdAt: string;
  /** @nullable */
  endDate: string | null;
  /** @nullable */
  startDate: string | null;
  /** @nullable */
  fixedValue: DiscountFixedValue;
  /** @nullable */
  percentage: DiscountPercentage;
  active: boolean;
  description: string;
  name: string;
  code: string;
  /** @nullable */
  organizationId: number | null;
  id: number;
}