// src/lib/generated/billing/models/updateDiscountDto.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * spaces-billing-and-subscription
 * Spaces Platform billing and subscription service
 * OpenAPI spec version: 1.0.0
 */
import type { UpdateDiscountDtoPercentage } from './updateDiscountDtoPercentage';
import type { UpdateDiscountDtoFixedValue } from './updateDiscountDtoFixedValue';

export interface UpdateDiscountDto {
  code?: string;
  name?: string;
  description?: string;
  active?: boolean;
  /** @nullable */
  percentage?: UpdateDiscountDtoPercentage;
  /** @nullable */
  fixedValue?: UpdateDiscountDtoFixedValue;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
}